提交说明： 1. 实现 或门 的基本逻辑 -- LogicElectricalAppliances/gate_circuit.GateOr 2. 实现 非门 的基本逻辑 -- LogicElectricalAppliances/gate_circuit.GateNot 3. 实现 与非门 的基本逻辑 -- LogicElectricalAppliances/gate_circuit.GateNon 4. 实现 或非门 的基本逻辑 -- LogicElectricalAppliances/gate_circuit.GateNor 5. 实现 异或门 的基本逻辑 -- LogicElectricalAppliances/gate_circuit.GateXor 1. 实现 同或门 的基本逻辑 -- LogicElectricalAppliances/gate_circuit.GateXnor

+++ LogicElectricalAppliances/src/gate_circuit/Gate.java
@@ -11,13 +11,13 @@ public interface Gate {
      * 作用：获取 <b>门</b> 输入端的值
      * @return 输入值
      */
-    int[] getInputValues();
+    byte[] getInputValues();
 
     /**
      * 方法名：获取输出端的值
      * 作用：获取 <b>门</b> 输出端的值
      * @return 输出值
      */
-    int getOutputValue();
+    byte getOutputValue();
 
 }
